HBAPP-176 Move components to repective folders
