+++ components/core/tailwind-indicator.tsx
