+++ components/features/forms/form-login.tsx
@@ -5,7 +5,7 @@ import Link from 'next/link'
 import { useTheme } from 'next-themes'
 import { ANON } from '@/lib/auth/providers/anonymous'
 import { UserInput } from '@/components/features/forms/input-username'
-import { LoginButton } from '@/components/login-button'
+import { LoginButton } from '@/components/features/forms/login-button'
 import { IconUsers, IconGitHub, IconVercel } from '@/components/ui/icons'
 import { Separator } from '@/components/ui/separator'
 import { Button } from '@/components/ui/button'
