+++ components/features/layout/footer.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 
 import { cn } from '@/lib/utils'
-import { ExternalLink } from '@/components/external-link'
+import { ExternalLink } from '@/components/ui/external-link'
 
 export function FooterText({ className, ...props }: React.ComponentProps<'p'>) {
   return (
