+++ components/features/layout/menu-panel-triggers.tsx
@@ -3,9 +3,9 @@ import { useCallback, useEffect } from 'react'
 import { Session } from 'next-auth'
 import { Button } from '@/components/ui/button'
 import { Tooltip, TooltipContent, TooltipTrigger } from '@/components/ui/tooltip'
-import { CollectionsButton } from '@/components/features/panels/collections-panel-button'
+import { CollectionsButton } from '@/components/features/panels/collection-panel-button'
 import { ChatsButton } from '@/components/features/panels/chats-panel-button'
-import { DocumentsButton } from '@/components/features/panels/documents-panel-button'
+import { DocumentsButton } from '@/components/features/panels/document-panel-button'
 import { ROUTE_CHATBOT, ROUTE_KNOWLEDGE, ROUTE_PLAYGROUND } from '@/app/constants'
 import { IconArrowElbow } from '@/components/ui/icons'
 import { CubeIcon } from '@radix-ui/react-icons'
