+++ components/features/menus/tabs/tab-knowledge.tsx
@@ -12,7 +12,7 @@ import { I_Collection, I_Knowledge_State, T_Memory_Type } from '@/lib/homebrew'
 import { Separator } from '@/components/ui/separator'
 import { Root, Indicator } from '@radix-ui/react-checkbox'
 import { CheckIcon } from '@radix-ui/react-icons'
-import { CollectionCard } from '@/components/sidebar-item-brain'
+import { CollectionCard } from '@/components/features/panels/collection-card'
 import ToggleGroup from '@/components/ui/toggle-group'
 import { IconBrain, IconDocument } from '@/components/ui/icons'
 import { DEFAULT_TYPE } from '@/components/features/menus/charm/hook-charm-knowledge'
