+++ components/features/panels/chats-panel-button.tsx
@@ -3,8 +3,8 @@
 import { Suspense } from 'react'
 import { Session } from 'next-auth/types'
 import { SidebarChatList } from '@/components/sidebar-list-chat'
-import { Sidebar } from '@/components/sidebar'
-import { SidebarFooter } from '@/components/sidebar-footer'
+import { Panel } from '@/components/features/panels/panel'
+import { PanelFooter } from '@/components/features/panels/panel-footer'
 import { ClearData } from '@/components/features/crud/dialog-clear-data'
 import { IconChat } from '@/components/ui/icons'
 
@@ -12,13 +12,13 @@ export const ChatsButton = ({ session }: { session: Session }) => {
   const clearChats = async () => true
 
   return (
-    <Sidebar title="Chat History" icon={IconChat}>
+    <Panel title="Chat History" icon={IconChat}>
       <Suspense fallback={<div className="flex-1 overflow-auto" />}>
         <SidebarChatList userId={session?.user?.id} />
       </Suspense>
-      <SidebarFooter>
+      <PanelFooter>
         <ClearData action={clearChats} actionTitle="Clear history" />
-      </SidebarFooter>
-    </Sidebar>
+      </PanelFooter>
+    </Panel>
   )
 }
